+++ product-service/src/main/java/com/stschool/ecommerce/productservice/projection/CategoryAveragePrice.java
@@ -1,8 +1,6 @@
 package com.stschool.ecommerce.productservice.projection;
 
-import java.math.BigDecimal;
-
 public interface CategoryAveragePrice {
     String getCategory();
-    BigDecimal getAvgPrice();
+    Double getAvgPrice();
 }
